Testing to see what all gets committed.
I made a change in .gitignore, but it isn't staged for commit so I don't expect anything to come through for that.

I made a change to using_api.py and then added that to the index.
I made _another_ change to using_api.py after it was added to the index. Think I'll see that in my commit?